feat: File Fetching | File Filtering
Two helper functions have been created:

The first one `fetch_files` visits a directory and fetches all
file and directory names, sorts them in 2 vectors (files and directories).

The second function `filter_files` takes vectors of directories in the
backup and main folder and a filter which is applied to them. If the
main folder contains an ignored folder, it simply is removed from the
vector, however if it exists in the backup folder, the folder [ignored one]
itself is removed.